+++ engine/src/main/java/dev/gamekit/ui/enums/CrossAxisAlignment.java
@@ -1,0 +1,16 @@
+package dev.gamekit.ui.enums;
+
+/**
+ * Determines how children of a {@link dev.gamekit.ui.widgets.Flex Flex}
+ * parent are aligned along the cross axis
+ */
+public enum CrossAxisAlignment {
+  /** Aligns all children to the start of the cross axis */
+  START,
+  /** Centers all children along the cross axis */
+  CENTER,
+  /** Aligns all children to the end of the cross axis */
+  END,
+  /** Stretches all children to same size on the cross axis */
+  STRETCH
+}