+++ engine/src/main/java/dev/gamekit/ui/enums/ImageFit.java
@@ -1,0 +1,11 @@
+package dev.gamekit.ui.enums;
+
+/** Determines how the image should be resized/fitted in its bounds */
+public enum ImageFit {
+  /** Resize the image to fit within the bounds */
+  FIT,
+  /** Cutout the portions of the image which are outside the bounds */
+  CROP,
+  /** Stretch the image to completely cover the bounds */
+  STRETCH
+}