+++ engine/src/main/java/dev/gamekit/ui/enums/TextAlignment.java
@@ -1,0 +1,14 @@
+package dev.gamekit.ui.enums;
+
+/**
+ * Determines the text alignment of a
+ * {@link dev.gamekit.ui.widgets.Text Text} widget
+ */
+public enum TextAlignment {
+  /** Left aligns the text content */
+  START,
+  /** Centers the text content */
+  CENTER,
+  /** Right aligns the text content */
+  END,
+}