+++ engine/src/main/java/dev/gamekit/ui/widgets/Center.java
@@ -1,6 +1,6 @@
 package dev.gamekit.ui.widgets;
 
-import dev.gamekit.ui.Alignment;
+import dev.gamekit.ui.enums.Alignment;
 
 /** A {@link Parent} which centers its single child */
 public class Center extends Align {